Shuffle some stuff.
