+++ moq-karp/Cargo.toml
@@ -44,8 +44,4 @@ features = ["add", "add_assign", "mul", "mul_assign", "sum", "from", "display"]
 
 [features]
 cli = ["moq-native", "tokio/full", "clap", "anyhow"]
-default = []
-
-[[bin]]
-name = "moq-karp"
-required-features = ["cli"]
+default = ["cli"]
