+++ moq-karp/src/lib.rs
@@ -1,22 +1,18 @@
 mod audio;
 mod broadcast;
 mod catalog;
-mod dimensions;
 mod error;
 mod frame;
-mod room;
-mod timestamp;
+mod member;
 mod track;
 mod video;
 
 pub use audio::*;
 pub use broadcast::*;
 pub use catalog::*;
-pub use dimensions::*;
 pub use error::*;
 pub use frame::*;
-pub use room::*;
-pub use timestamp::*;
+pub use member::*;
 pub use track::*;
 pub use video::*;
