+++ moq-karp/src/main.rs
@@ -4,7 +4,7 @@ use anyhow::Context;
 use clap::{Parser, Subcommand};
 use url::Url;
 
-use moq_karp::{cmaf, produce};
+use moq_karp::{cmaf, Member};
 use moq_native::quic;
 
 #[derive(Parser, Clone)]
@@ -21,17 +21,13 @@ struct Cli {
 	#[command(flatten)]
 	pub tls: moq_native::tls::Args,
 
-	/// Connect to the given URL starting with https://
-	#[arg(long, default_value = "https://relay.quic.video")]
-	pub url: Url,
-
-	/// The path of the broadcast
-	/// Use multiple times to create a nested path.
-	#[arg(long, required = true)]
-	pub path: Vec<String>,
-
+	/// If we're publishing or subscribing.
 	#[command(subcommand)]
 	pub command: Command,
+
+	/// The broadcast address.
+	#[arg()]
+	pub addr: Url,
 }
 
 #[derive(Subcommand, Clone)]
@@ -48,11 +44,11 @@ async fn main() -> anyhow::Result<()> {
 	let tls = cli.tls.load()?;
 	let quic = quic::Endpoint::new(quic::Config { bind: cli.bind, tls })?;
 
-	tracing::info!(url = %cli.url, "connecting");
-	let session = quic.client.connect(&cli.url).await?;
-	let session = moq_transfork::Session::connect(session).await?;
+	let path = moq_transfork::Path::from_iter(cli.addr.path_segments().context("no broadcast path")?);
 
-	let path = moq_transfork::Path::new(cli.path);
+	tracing::info!(url = %cli.addr, "connecting");
+	let session = quic.client.connect(&cli.addr).await?;
+	let session = moq_transfork::Session::connect(session).await?;
 
 	match cli.command {
 		Command::Publish => publish(session, path).await,
@@ -61,17 +57,15 @@ async fn main() -> anyhow::Result<()> {
 }
 
 #[tracing::instrument("publish", skip_all, err, fields(?path))]
-async fn publish(mut session: moq_transfork::Session, path: moq_transfork::Path) -> anyhow::Result<()> {
-	let broadcast = produce::Resumable::new(path).broadcast();
+async fn publish(session: moq_transfork::Session, path: moq_transfork::Path) -> anyhow::Result<()> {
+	let user = Member::new(session.clone(), path).produce().broadcast_now()?;
 
 	let mut input = tokio::io::stdin();
 
-	let mut import = cmaf::Import::new(broadcast);
+	let mut import = cmaf::Import::new(user);
 	import.init_from(&mut input).await.context("failed to initialize")?;
 
 	tracing::info!(catalog = ?import.catalog());
-
-	import.publish(&mut session)?;
 
 	tokio::select! {
 		res = import.read_from(&mut input) => Ok(res?),
